feat: implement hover state and program management features
Eliminate [[maybe_unused]] annotations by implementing actual functionality:

1. PluginProcessor program management:
   - Track current program index with member variable
   - Implement getProgramName() to return "Sidechain" for program 0
   - Implement setCurrentProgram() with bounds checking and logging
   - Implement changeProgramName() with debug logging (not supported)

2. UI hover state effects:
   - Comment.cpp: Implement conditional hover effect rendering
   - NotificationBell.cpp: Implement notification bell hover effects
   - NotificationList.cpp: Implement notification row hover effects
   - FollowersList.cpp: Implement follow button hover effects
   - UserCard.cpp: Implement user card elevation and button hover

3. Animation delay scheduling:
   - AnimationTimeline.h: Properly use delayMs parameter for stagger offsets
   - Adjust start time based on delay for sequential animations

Build: 0 errors, [[maybe_unused]] count: 55 remaining

+++ plugin/src/util/cache/FileCache.cpp
@@ -277,7 +277,7 @@ FileCache::CacheEntry FileCache::CacheEntry::fromJSON(const juce::var &json) {
   CacheEntry entry;
   entry.url = json.getProperty("url", "").toString();
   entry.filename = json.getProperty("filename", "").toString();
-  entry.fileSize = static_cast<int64_t>(json.getProperty("fileSize", 0.0));
+  entry.fileSize = static_cast<int64_t>(static_cast<double>(json.getProperty("fileSize", 0.0)));
   entry.lastAccessTime = json.getProperty("lastAccessTime", 0.0);
 
   if (entry.url.isEmpty() || entry.filename.isEmpty()) {
